extract user list to fragment
