+++ app/build.gradle
@@ -52,6 +52,8 @@ dependencies {
     kapt "com.squareup.moshi:moshi-kotlin-codegen:$moshiVersion"
 
     implementation "androidx.activity:activity-ktx:1.4.0"
+    implementation "androidx.fragment:fragment-ktx:1.4.0-rc01"
+
     implementation "com.github.kirich1409:viewbindingpropertydelegate-noreflection:1.4.4"
 
     implementation 'androidx.core:core-ktx:1.7.0'
