+++ app/src/main/java/com/example/myapplication/UserListViewModel.kt
@@ -14,7 +14,7 @@ import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 
-class MainViewModel : ViewModel() {
+class UserListViewModel : ViewModel() {
     sealed class ViewState {
         object Loading : ViewState()
         data class Data(val users: List<User>) : ViewState()
